+++ History.md
@@ -1,3 +1,11 @@
+0.1.6 / 2011-02-26 
+==================
+
+  * Attribute inheritance/merging with `merge-attrs=""` for all, `merge-attrs="name"` for named attributes
+  * Changed `callback` to `tagbody` to use in by `tagbody.call(this)` in EJS
+  * Better express error integration
+  * Added underscore library to use in views
+  
 0.1.5 / 2011-02-23 
 ==================